requirements.txt
fastapi
uvicorn[standard]
httpx
Pillow
numpy
voyageai
aiosqlite
aiofiles
pynput # For idle detection
python-dotenv # For .env file
pyobjc; sys_platform == 'darwin'  # Required for Apple Vision on macOS
pandas>=1.3.0
scikit-learn>=1.0.0
torch>=1.9.0
plotly>=5.3.0
# For macOS specific metadata:
